New built-in function lbound_of(s-att, _) and rbound_of(s-att, _) to determine start and end of region containing corpus position (_ or label reference).
Built-in functions called with undefined values (ATTAT_NONE) now return ATTAT_NONE rather than
raising an error. This was necessitated by the new lbound_of() and rbound_of() functions, which
return ATTAT_NONE outside a suitable s-attribute region.

Bug fix: bare s-attribute reference returns code 0 outside a suitable region, as documented in
the source code; previously, it raised an evaluation error that caused all comparisons and 
function calls to evaluate to false (which was the expected behaviour in most contexts, but 
prevented the implementation of lbound_of() and rbound_of())




git-svn-id: https://svn.code.sf.net/p/cwb/code/cwb/trunk@1037 b72dbb9a-5b02-45d8-a000-b69eecba7de2